# src/main/resources/openapi/pms-api.yaml
openapi: 3.0.3
info:
  title: PMS 项目管理 API
  version: 1.0.0
  description: 单企业私有化部署的项目、组织和权限管理接口。
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api
...
paths:
...
  /projects/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [项目]
      summary: 项目详情
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [项目]
      summary: 更新项目
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                priority: { type: integer, minimum: 0, maximum: 3, description: 低、中、高、紧急 }
                projectLevel: { type: integer, enum: [0, 1, 2, 3], description: 0常规(C)、1重要(B)、2关键(A)、3战略(S) }
                orgUnitId: { type: integer, format: int64 }
                projectManagerId: { type: integer, format: int64, nullable: true }
                memberIds: { type: array, items: { type: integer, format: int64 } }
                expectedMemberIds: { type: array, items: { type: integer, format: int64 }, description: 客户端上次看到的成员集合，用于避免覆盖并发自动加入的成员 }
                followerIds: { type: array, items: { type: integer, format: int64 } }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [项目]
      summary: 删除项目
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /projects/{projectId}/images:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [项目]
      summary: 上传项目图片
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /projects/{projectId}/images/{filename}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: filename
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [项目]
      summary: 读取项目图片
      responses:
        '200': { description: 图片二进制 }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [项目]
      summary: 删除项目图片
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        msg: { type: string }
        data: {}
        requestId: { type: string }
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required: [code, msg, requestId]
  responses:
    Success:
      description: 请求成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
...
    Forbidden:
      description: 无权访问
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 数据冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    TooManyRequests:
      description: 请求过于频繁
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
security:
  - bearerAuth: []